requirements.txt
opencv-python>=4.9.0.80
tensorflow-cpu==2.15.0
flask==3.0.3
flask-cors==4.0.1
werkzeug==3.1.3
numpy==1.26.4
pillow==10.3.0
requests==2.31.0
streamlit
